+++ src/Middleware/MiddlewareIntegration.php
@@ -38,6 +38,8 @@ class MiddlewareIntegration implements \Rougin\Slytherin\Integration\Integration
             $dispatcher->push($middleware);
         }
 
+        // NOTE: To be removed in v1.0.0. Use Middleware\DispatcherInterface instead.
+        $container->set('Rougin\Slytherin\Middleware\MiddlewareInterface', $dispatcher);
         $container->set('Rougin\Slytherin\Middleware\DispatcherInterface', $dispatcher);
 
         return $container;